+++ Dalamud.Boot/pch.h
@@ -8,7 +8,8 @@
 #define PCH_H
 
 // Exclude rarely-used stuff from Windows headers
-#define WIN32_LEAN_AND_MEAN 
+#define WIN32_LEAN_AND_MEAN
+#define NOMINMAX
 
 // Windows Header Files
 #include <windows.h>
